+++ src/storage/src/test_util.rs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+pub mod access_layer_util;
 pub mod config_util;
 pub mod descriptor_util;
 pub mod read_util;
